moving functions to lib
