+++ lib/api/mapBenefits.js
@@ -1,11 +1,15 @@
-import { CPPPrograms, CPPStatus, CPPTypes } from '../constants/CPPMappingCodes'
-import { EIStatus, EITypes } from '../constants/EIMappingCodes'
-import { OASStatus, OASTypes } from '../constants/OASMappingCodes'
-import { ProgramCodes } from '../constants/ProgramCodes'
-import { TypeCodes } from '../constants/ProgramTypeCodes'
-import { StatusCodes } from '../constants/StatusCodes'
-import { SummaryTypes } from '../constants/SummaryTypes'
-import { TASK_GROUPS } from '../contents/DashboardBenefitTasksConstants'
+import {
+  CPPPrograms,
+  CPPStatus,
+  CPPTypes,
+} from '../../constants/CPPMappingCodes'
+import { EIStatus, EITypes } from '../../constants/EIMappingCodes'
+import { OASStatus, OASTypes } from '../../constants/OASMappingCodes'
+import { ProgramCodes } from '../../constants/ProgramCodes'
+import { TypeCodes } from '../../constants/ProgramTypeCodes'
+import { StatusCodes } from '../../constants/StatusCodes'
+import { SummaryTypes } from '../../constants/SummaryTypes'
+import { TASK_GROUPS } from '../../contents/DashboardBenefitTasksConstants'
 
 function MapBenefitToTaskGroups(programCode, statusCode) {
   const group = TASK_GROUPS.find(
@@ -191,7 +195,7 @@ export function MapSEBCard(sebResult) {
   )
   return benefit
 }
-  
+
 ///OAS result object maping
 // {
 //     "programCode": oas,
