+++ pages/api/programData/cpp.js
@@ -1,4 +1,4 @@
-import { MapCPPCard } from '../../../lib/BenefitsMapping'
+import { MapCPPCard } from '../../../lib/api/mapBenefits'
 import { GetProgramData } from './_middleware'
 
 export default async function handler(req, res) {
@@ -6,7 +6,7 @@ export default async function handler(req, res) {
     req,
     res,
     process.env.CPP_ACTIVE_BENEFIT_URL,
-    (mockObj) => mockObj.CPP,
-    (data) => MapCPPCard(data)
+    (data) => MapArrayData(data, (item) => MapCPPCard(item)),
+    (mockObj) => mockObj.CPP
   )
 }
